build: support DTrace on Linux
This change allows the builder to use either SystemTap or DTrace on
Linux, by introducing `--dtrace-mode` flag to `configure.py`. Valid
options include `auto` (default), `traditional` and `mac`.

For auto on Linux, the check will probe for `/usr/sbin/dtrace` first,
and if it exists assume the user wants "traditional" DTrace, then it
will probe for `/usr/bin/dtrace` and assume the user wants SystemTap,
if neither file exists it will warn and assume the original behavior of
SystemTap.

For platforms that aren't Linux, the behavior should remain the same.